refactor: add graph page
