+++ expense-app/src/components/Tabs/GraphTab.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+function GraphTab() {
+  return (
+    <div>GraphTab</div>
+  )
+}
+
+export default GraphTab
