+++ expense-app/src/components/Tabs/tab.scss
