+++ expense-app/src/pages/home/Home.scss
@@ -1,10 +1,6 @@
 @import './../../styles/index.scss';
 
-.home {
-    margin-left: 200px;
-    background: rgb(243, 235, 235);
-    height: 100vh;
-}
+
 .wrapper {
     position: relative;
     display: block;
