+++ expense-app/src/styles/index.scss
@@ -12,3 +12,9 @@ code {
   font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New',
     monospace;
 }
+
+.page-wrapper {
+  margin-left: 200px;
+  background: rgb(243, 235, 235);
+  height: 100vh;
+}
